requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
anthropic==0.18.1
pydub==0.25.1
pyaudio==0.2.14
pyannote.audio==4.0.1
torch==2.1.0
torchaudio==2.1.0
python-docx==1.1.0
reportlab==4.0.7
requests==2.31.0
openai==1.12.0
numpy==1.26.3
scipy==1.12.0
librosa==0.10.2
soundfile==0.12.1
ffmpeg-python==0.2.0
Pillow==10.2.0
gunicorn==21.2.0

